+++ docs/release-notes.md
@@ -1,2 +1,4 @@
+# May 5 2017
+ - Add platform specific commands and timers
 # April 2017
  - Command Aliases